Added auth in resolvers

+++ Intro to GraphQL/src/types/product/product.resolvers.js
@@ -10,19 +10,34 @@ const productsTypeMatcher = {
   DRONE: 'Drone'
 }
 
-const products = () => {
+const products = (_,args,ctx) => {
+  if(!ctx.user){
+    throw new AuthenticationError
+  }
   return Product.find({}).exec()
 }
-const product  = (_,args) => {
+const product  = (_,args,ctx) => {
+  if(!ctx.user){
+    throw new AuthenticationError
+  }
   return Product.findById({_id:args.id}).exec()
 }
 const newProduct = (_,args,ctx) =>{
+  if(!ctx.user || ctx.user.role != 'admin'){
+    throw new AuthenticationError
+  }
   return Product.create({...args.input, createdBy: ctx.user._id})
 }
-const updateProduct = (_, args) =>{
-  return Product.findByIdAndUpdate(args.id , args.input, {new:true}).lean().exec()
+const updateProduct = (_, args,ctx) =>{
+  if(!ctx.user || ctx.user.role != 'admin'){
+    throw new AuthenticationError
+  }
+  return Product.findByIdAndUpdate(args.id , args.input, {new:true}).exec()
 }
-const removeProduct = (_,args) => {
+const removeProduct = (_,args,ctx) => {
+  if(!ctx.user || ctx.user.role != 'admin'){
+    throw new AuthenticationError
+  }
   return Product.findByIdAndDelete(args.id).lean().exec()
 }
 export default {
